+++ scrapli_netconf/channel/async_channel.py
@@ -26,14 +26,14 @@ class AsyncNetconfChannel(AsyncChannel, NetconfChannelBase):
         At least per early drafts of the netconf over ssh rfcs the netconf servers MUST NOT echo the
         input commands back to the client. In the case of "normal" scrapli netconf with the system
         transport this happens anyway because we combine the stdin and stdout fds into a single pty,
-        however for asyncssh we have an actual stdin and stdout fd to read/write. It seems that at
-        the very least Juniper seems to want to echo inputs back onto to the stdout for the channel.
-        This is totally ok and we can deal with it, we just need to *know* that it is happening and
-        that gives us somewhat of a dilemma... we want to give the device time to echo this data
-        back to us, but we also dont want to just arbitrarily wait (especially in the more common
-        case where the device is *not* echoing anything back). So we take 1/20th of the transport
-        timeout and we wait that long to see -- if we get echo, we return immediately of course,
-        otherwise there is an unfortunate slight delay here :(
+        however for other transports we have an actual stdin and stdout fd to read/write. It seems
+        that at the very least IOSXE with NETCONF 1.1 seems to want to echo inputs back onto to the
+        stdout for the channel. This is totally ok and we can deal with it, we just need to *know*
+        that it is happening and that gives us somewhat of a dilemma... we want to give the device
+        time to echo this data back to us, but we also dont want to just arbitrarily wait
+        (especially in the more common case where the device is *not* echoing anything back). So we
+        take 1/20th of the transport timeout and we wait that long to see -- if we get echo, we
+        return immediately of course, otherwise there is an unfortunate slight delay here :(
 
         See: https://tools.ietf.org/html/draft-ietf-netconf-ssh-02 (search for "echo")
 
@@ -171,10 +171,16 @@ class AsyncNetconfChannel(AsyncChannel, NetconfChannelBase):
 
         """
         final_channel_input = self._build_message(channel_input)
+        bytes_final_channel_input = final_channel_input.encode()
 
         raw_result, _ = await super().send_input(
-            channel_input=final_channel_input, strip_prompt=False
+            channel_input=final_channel_input, strip_prompt=False, eager=True
         )
+
+        if bytes_final_channel_input in raw_result:
+            raw_result = raw_result.split(bytes_final_channel_input)[1]
+
+        raw_result = await self._read_until_prompt(output=raw_result)
 
         if self.netconf_version == NetconfVersion.VERSION_1_1:
             # netconf 1.1 with "chunking" style message format needs an extra return char here