+++ scrapli_netconf/response.py
@@ -2,7 +2,7 @@
 import logging
 import re
 from datetime import datetime
-from typing import Any, Dict, List, Optional, Union
+from typing import Any, Dict, List, Optional, Tuple, Union
 
 from lxml import etree
 from lxml.etree import Element
@@ -15,8 +15,12 @@ LOG = logging.getLogger("response")
 
 # "chunk match" matches two groups per section returned from the netconf server, first the length of
 # the response, and second the response itself. we use the length of the response to validate the
-# response is in fact X length
-CHUNK_MATCH_1_1 = re.compile(pattern=rb"^#(\d+)(?:\n*)(((?!#).)*)", flags=re.M | re.S)
+# response is in fact X length. this regex is basically "start at line feed, and match "#123" where
+# "123" is obviously any length of digits... then we don't capture zero or more newlines because we
+# dont care about them. Next we have the main capture group -- this starts with a negative lookahead
+# that says we want to stop matching as soon as we hit another "#123" *or* a "##" (end of message),
+# after that we match anything "." and that is the "body" of the response
+CHUNK_MATCH_1_1 = re.compile(pattern=rb"^#(\d+)(?:\n*)(((?!#\d+\n+|##).)*)", flags=re.M | re.S)
 
 PARSER = etree.XMLParser(remove_blank_text=True)
 
@@ -27,7 +31,7 @@ class NetconfResponse(Response):
         netconf_version: NetconfVersion,
         xml_input: Element,
         strip_namespaces: bool = True,
-        failed_when_contains: Optional[Union[bytes, List[bytes]]] = b"<rpc-error>",
+        failed_when_contains: Optional[Union[bytes, List[bytes]]] = None,
         **kwargs: Any,
     ):
         """
@@ -63,9 +67,20 @@ class NetconfResponse(Response):
 
         super().__init__(**kwargs)
 
+        if failed_when_contains is None:
+            # match on both opening and closing tags too so we never have to think about/compare
+            # things with namespaces (the closing tags wont have namespaces)
+            failed_when_contains = [
+                b"</rpc-error>",
+                b"</rpc-errors>",
+                b"<rpc-error>",
+                b"<rpc-errors>",
+            ]
         if isinstance(failed_when_contains, bytes):
             failed_when_contains = [failed_when_contains]
         self.failed_when_contains = failed_when_contains
+
+        self.error_messages: List[str] = []
 
     def _record_response(self, result: bytes) -> None:
         """
@@ -85,10 +100,18 @@ class NetconfResponse(Response):
         self.elapsed_time = (self.finish_time - self.start_time).total_seconds()
         self.raw_result = result
 
+        if not self.failed_when_contains:
+            self.failed = False
+        elif not any(err in self.raw_result for err in self.failed_when_contains):
+            self.failed = False
+
         if self.netconf_version == NetconfVersion.VERSION_1_0:
             self._record_response_netconf_1_0()
         else:
             self._record_response_netconf_1_1()
+
+        if self.failed:
+            self._fetch_error_messages()
 
     def _record_response_netconf_1_0(self) -> None:
         """
@@ -104,11 +127,6 @@ class NetconfResponse(Response):
             N/A
 
         """
-        if not self.failed_when_contains:
-            self.failed = False
-        elif not any(err in self.raw_result for err in self.failed_when_contains):
-            self.failed = False
-
         # remove the message end characters and xml document header see:
         # https://github.com/scrapli/scrapli_netconf/issues/1
         self.xml_result = etree.fromstring(
@@ -124,6 +142,83 @@ class NetconfResponse(Response):
         else:
             self.result = etree.tostring(self.xml_result, pretty_print=True).decode()
 
+    def _validate_chunk_size_netconf_1_1(self, result: Tuple[str, bytes]) -> None:
+        """
+        Validate individual chunk size; handle parsing trailing new lines for chunk sizes
+
+        It seems that some platforms behave slightly differently than others (looking at you IOSXE)
+        in the way they count chunk sizes with respect to trailing whitespace. Per my reading of the
+        RFC, the response for a netconf 1.1 response should look like this:
+
+        ```
+        ##XYZ
+        <somexml>
+        ##
+        ```
+
+        Where "XYZ" is an integer number of the count of chars in the following chunk (the chars up
+        to the next "##" symbols), then the actual XML response, then a new line(!!!!) and a pair of
+        hash symbols to indicate the chunk is complete.
+
+        IOSXE seems to *not* want to see the newline between the XML payload and the double hash
+        symbols... instead when it sees that newline it immediately returns the response. This
+        breaks the core behavior of scrapli in that scrapli always writes the input, then reads the
+        written inputs off the channel *before* sending a return character. This ensures that we
+        never have to deal with stripping out the inputs and such because it has already been read.
+        With IOSXE Behaving this way, we have to instead use `send_input` with the `eager` flag set
+        -- this means that we do *not* read the inputs, we simply send a return. We then have to do
+        a little extra parsing to strip out the inputs, but thats no big deal...
+
+        Where this finally gets to "spacing" -- IOSXE seems to include trailing newlines *sometimes*
+        but not other times, whereas IOSXR (for example) *always* counts a single trailing newline
+        (after the XML). SO.... long story long... (the above chunk stuff doesn't necessarily matter
+        for this, but felt like as good a place to document it as any...) this method deals w/
+        newline counts -- we check the expected chunk length against the actual char count, the char
+        count with all trailing whitespace stripped, and the count of the chunk + a *single*
+        trailing newline character...
+
+        FIN
+
+        Args:
+            result: Tuple from re.findall parsing the full response object
+
+        Returns:
+            N/A  # noqa: DAR202
+
+        Raises:
+            N/A
+
+        """
+        expected_len = int(result[0])
+        result_value = result[1]
+
+        actual_len = len(result_value)
+        rstripped_len = len(result_value.rstrip())
+
+        trailing_newline_count = actual_len - rstripped_len
+        if trailing_newline_count > 1:
+            extraneous_trailing_newline_count = trailing_newline_count - 1
+        else:
+            extraneous_trailing_newline_count = 1
+        trimmed_newline_len = actual_len - extraneous_trailing_newline_count
+
+        if expected_len == 1:
+            # at least nokia tends to have itty bitty chunks of one element, deal w/ that
+            actual_len = 1
+
+        if expected_len == actual_len:
+            return
+        if expected_len == rstripped_len:
+            return
+        if expected_len == trimmed_newline_len:
+            return
+
+        LOG.critical(
+            f"Return element length invalid, expected {expected_len} got {actual_len} for "
+            f"element: {repr(result_value)}"
+        )
+        self.failed = True
+
     def _record_response_netconf_1_1(self) -> None:
         """
         Record response for netconf version 1.1
@@ -138,29 +233,12 @@ class NetconfResponse(Response):
             N/A
 
         """
-        if not self.failed_when_contains:
-            self.failed = False
-        elif not any(err in self.raw_result for err in self.failed_when_contains):
-            self.failed = False
-
         result_sections = re.findall(pattern=CHUNK_MATCH_1_1, string=self.raw_result)
 
         # validate all received data
         for result in result_sections:
-            expected_len = int(result[0])
-            result_value = result[1]
-            # account for trailing newline char
-            actual_len = len(result_value) - 1
-            if expected_len == 1:
-                # at least nokia tends to have itty bitty chunks of one element, deal w/ that
-                actual_len = 1
-            if expected_len != actual_len:
-                LOG.critical(
-                    f"Return element length invalid, expected {expected_len} got {actual_len} for "
-                    f"element: {repr(result_value)}"
-                )
-                self.failed = True
-
+            self._validate_chunk_size_netconf_1_1(result=result)
+        # assert 0
         self.xml_result = etree.fromstring(
             b"\n".join(
                 [
@@ -178,6 +256,27 @@ class NetconfResponse(Response):
             self.result = etree.tostring(self.xml_result, pretty_print=True).decode()
         else:
             self.result = etree.tostring(self.xml_result, pretty_print=True).decode()
+
+    def _fetch_error_messages(self) -> None:
+        """
+        Fetch all error messages (if any)
+
+        RFC states that there MAY be more than one rpc-error so we just xpath for all
+        "error-message" tags and pull out the text of those elements. The strip is just to remove
+        leading/trailing white space to make things look a bit nicer.
+
+        Args:
+            N/A
+
+        Returns:
+            N/A  # noqa: DAR202
+
+        Raises:
+            N/A
+
+        """
+        err_messages = self.xml_result.xpath("//rpc-error/error-message")
+        self.error_messages = [err.text.strip() for err in err_messages]
 
     def get_xml_elements(self) -> Dict[str, Element]:
         """