+++ scrapli_netconf/transport/asyncssh_.py
@@ -3,6 +3,8 @@ from typing import Any
 
 from scrapli_asyncssh.transport.asyncssh_ import AsyncSSHTransport
 
+from scrapli.exceptions import ConnectionNotOpened
+
 
 # no stubs for scrapli_asyncssh for now, ignore strict typing complains
 class NetconfAsyncSSHTransport(AsyncSSHTransport):  # type: ignore
@@ -23,6 +25,8 @@ class NetconfAsyncSSHTransport(AsyncSSHTransport):  # type: ignore
             N/A
 
         """
+        from asyncssh.misc import ChannelOpenError  # pylint: disable=C0415
+
         if self.auth_strict_key:
             self.logger.debug(f"Attempting to validate {self.host} public key is in known hosts")
             self._verify_key()
@@ -38,9 +42,20 @@ class NetconfAsyncSSHTransport(AsyncSSHTransport):  # type: ignore
         # it seems we must pass a terminal type to force a pty(?) which i think we want in like...
         # every case?? https://invisible-island.net/ncurses/ncurses.faq.html#xterm_color
         # set encoding to None so we get bytes for consistency w/ other scrapli transports
-        self.stdin, self.stdout, self.stderr = await self.session.open_session(
-            term_type="xterm", encoding=None, subsystem="netconf"
-        )
+        # request_pty seems to be safe to set to "false" but leaving it at auto which seems closer
+        # to the default behavior. With this omitted (as was previously) connecting to junos devices
+        # on port 830 worked w/ system transport but *not* asyncssh because the pty request failed
+        try:
+            self.stdin, self.stdout, self.stderr = await self.session.open_session(
+                term_type="xterm", encoding=None, subsystem="netconf", request_pty="auto"
+            )
+        except ChannelOpenError as exc:
+            msg = (
+                "Failed to open Channel -- do you have the right port? Most often the netconf "
+                "port is 22 or 830!"
+            )
+            self.logger.critical(msg)
+            raise ConnectionNotOpened(msg) from exc
 
         return b""